Updated freeTubeUpdate.sh

+++ shl/freeTubeUpdate.sh
@@ -9,17 +9,22 @@ esac
 
 scriptDir=$(dirname $0)
 scriptDir=$(cd $scriptDir;pwd)
+
 app=freeTube
-gitHubURL=https://github.com
-gitHubAPIURL=https://api.github.com
+protocol=https
+gitHubURL=github.com
+gitHubAPIURL=api.$gitHubURL
 gitHubUser=FreeTubeApp
 gitHubRepo=FreeTube
 gitHubAPIRepoURL=$gitHubAPIURL/repos/$gitHubUser/$gitHubRepo
-freeTubeLatestGitHubReleaseURL=$(curl -s $gitHubAPIRepoURL/releases | jq -r ".[0].assets[] | select( .name | contains( \"$arch.deb\") ) | .browser_download_url")
-freeTubeLatestGitHubReleaseTag=$(curl -s $gitHubAPIRepoURL/releases | jq -r '.[0].tag_name' | sed 's/v//;s/-beta//')
+
+freeTubeLatestGitHubReleaseTag=$(\curl -Ls $protocol://$gitHubAPIRepoURL/releases | jq -r '.[0].tag_name' | sed 's/v//;s/-beta//')
 freeTubeInstalledVersion=$(dpkg-query --showformat='${Version}' -W freetube)
 if [ "$freeTubeLatestGitHubReleaseTag" != "$freeTubeInstalledVersion" ];then
-	freeTubeLatestGitHubReleaseName=$(basename $freeTubeLatestGitHubReleaseURL)
-	wget -O $freeTubeLatestGitHubReleaseName "$freeTubeLatestGitHubReleaseURL"
-	sudo gdebi -n $freeTubeLatestGitHubReleaseName && rm -v $freeTubeLatestGitHubReleaseName
+	freeTubeLatestGitHubReleaseURL=$(\curl -Ls $protocol://$gitHubAPIRepoURL/releases | jq -r ".[0].assets[] | select( .name | contains( \"$arch.deb\") ) | .browser_download_url")
+	if [ -n "$freeTubeLatestGitHubReleaseURL" ];then
+		freeTubeLatestGitHubReleaseName=$(basename $freeTubeLatestGitHubReleaseURL)
+		wget -O $protocol://$freeTubeLatestGitHubReleaseName "$freeTubeLatestGitHubReleaseURL"
+		sudo gdebi -n $freeTubeLatestGitHubReleaseName && rm -v $freeTubeLatestGitHubReleaseName
+	fi
 fi
